+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # `latest`
 
-* **Buying things by their names**
+* **Buy things by their names**
+* **Raid things by their names**
 * Switch back to plain old `requirements.txt`
 * Fix `--help`